refactor: move find_event_id_from_last_e_tag into domain layer
Extracting the event ID referenced by the last `e` tag is generic Nostr
tag parsing (used for reactions, reposts and zap receipts), not a
responsibility of the `Timeline` model. Keeping it there leaked NIP tag
structure knowledge into the timeline orchestration logic.

Move it to `domain::nostr` as a free function next to `SortableEventId`,
re-export it, and update `Timeline` to call it. The associated unit tests
move alongside the function. No behavior change.
