+++ src/domain/nostr/event.rs
@@ -47,6 +47,24 @@ impl Ord for SortableEventId {
     }
 }
 
+/// Extract the event ID referenced by the last `e` tag of an event.
+///
+/// Reactions (NIP-25), reposts (NIP-18) and zap receipts (NIP-57) reference
+/// their target event with an `e` tag; when multiple `e` tags are present, the
+/// last one identifies the target. Returns `None` if there is no `e` tag.
+pub fn find_event_id_from_last_e_tag(event: &Event) -> Option<EventId> {
+    event
+        .tags
+        .filter_standardized(TagKind::SingleLetter(SingleLetterTag::lowercase(
+            Alphabet::E,
+        )))
+        .last()
+        .and_then(|tag| match tag {
+            TagStandard::Event { event_id, .. } => Some(*event_id),
+            _ => None,
+        })
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -142,5 +160,47 @@ mod tests {
         // Should be sorted in reverse order (newest first)
         let sorted: Vec<_> = set.iter().map(|r| r.0).collect();
         assert_eq!(sorted, vec![id3, id2, id1]);
+    }
+
+    #[test]
+    fn test_find_event_id_from_last_e_tag() {
+        let keys = Keys::generate();
+        let target_id = EventId::all_zeros();
+
+        let event = EventBuilder::new(Kind::Reaction, "+")
+            .tags(vec![Tag::event(target_id)])
+            .sign_with_keys(&keys)
+            .expect("Failed to create event");
+
+        let found_id = find_event_id_from_last_e_tag(&event);
+        assert_eq!(found_id, Some(target_id));
+    }
+
+    #[test]
+    fn test_find_event_id_from_last_e_tag_multiple_tags() {
+        let keys = Keys::generate();
+        let first_id = EventId::all_zeros();
+        let last_id = EventId::from_slice(&[1u8; 32]).expect("Valid event ID");
+
+        let event = EventBuilder::new(Kind::Reaction, "+")
+            .tags(vec![Tag::event(first_id), Tag::event(last_id)])
+            .sign_with_keys(&keys)
+            .expect("Failed to create event");
+
+        // Should return the last 'e' tag
+        let found_id = find_event_id_from_last_e_tag(&event);
+        assert_eq!(found_id, Some(last_id));
+    }
+
+    #[test]
+    fn test_find_event_id_from_last_e_tag_no_tags() {
+        let keys = Keys::generate();
+
+        let event = EventBuilder::new(Kind::Reaction, "+")
+            .sign_with_keys(&keys)
+            .expect("Failed to create event");
+
+        let found_id = find_event_id_from_last_e_tag(&event);
+        assert_eq!(found_id, None);
     }
 }
